+++ packages/ethereum-contracts/CHANGELOG.md
@@ -5,9 +5,14 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Unreleased
 
+### Added
+- `increaseFlowRateAllowanceWithPermissions` and `decreaseFlowRateAllowanceWithPermissions` added to `ConstantFlowAgreementV1.sol`
+
 ### Changed
 - `SuperToken.sol` made external and public methods virtual to facilitate creation of customized implementations.
 - Explicitly set EVM target to "paris" because EIP-3855 isn't yet supported on all chains with Superfluid deployment.
+- `SuperfluidGovernanceBase`: add a method for updating SuperTokens to other than the canonical implementation.
+- `SuperToken.sol` invoke mint/burn hooks only if userData is not empty
 
 ## [v1.7.1] - 2023-06-27